Add configurable response headers
A request was made to allow for custom HTTP headers to be returned from
the server per route.

This adds a new configuraiton option that is a map of header keys and
values. Thesea are then mapped to the response as required.

Also made some improvements to the http server, configuration and docs